+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 CHANGELOG jbrotli
 =================
 
+## v0.3.1 (2016-??-??)
+
+* NEW: provide BrotliServletFilter, so you may add to your Tomcat, Jetty, etc.
+* NEW: Brotli.Parameter has fluent setter-methods, for convenience
+* CHG: rename BrotliStreamCompressor array methods to 'compressArray()'
+
 ## v0.3.0 (2016-01-27)
 
 * NEW: update using brotli release 0.3.0